Left-recursion demo
Parser combinators remain my parsing strategy of choice because of the
simplicity of writing a new parser, but they have the notable drawback
of being useless for recursive descent parsing. Loco will detect
left-recursion at Grammar instantiation time. A way around this would be
to offer a dedicated LeftRecursionParser class, but until that time,
here's the quick hack.