Fix BNF failure
This thing was throwing a ParseFailureException. The problem was that it
was set up to only recognise "\n" as a line ending in the input string.
Solution: recognise either "\r" or "\n" as a line ending. (That means
"\r\n" counts as two line endings, but in this particular case we don't
care about empty lines so no big deal.)